requirements.txt
# Core & Data Handling
numpy==2.3.3
pandas==2.3.3
pyarrow==21.0.0
altair==5.5.0
pydeck==0.9.1

# Visualization & UI
streamlit==1.50.0
streamlit-lottie==0.0.5
Pillow==11.3.0

# Google AI & APIs
google-generativeai==0.8.5
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-ai-generativelanguage==0.6.15
grpcio==1.75.1
grpcio-status==1.71.2
proto-plus==1.26.1
protobuf==5.29.5

# Utilities
requests==2.32.3
python-dotenv==1.0.0
tenacity==9.1.2
tqdm==4.66.5

# Validation & Typing
pydantic==2.10.3
pydantic-settings==2.6.1
typing-extensions==4.12.2

# Auth & Security
cryptography==43.0.3
PyJWT==2.9.0
rsa==4.9.1

# CLI & Logging
click==8.1.7
typer==0.12.5
rich==13.9.2
colorama==0.4.6
watchdog==6.0.0
